+++ modules/__tests__/createFetch-test.js
@@ -1,13 +1,14 @@
 import expect from 'expect'
 import {
-  fetch as globalFetch,
   createFetch
 } from '../index'
+
+const global = (1, eval)('this')
 
 describe('createFetch', () => {
   describe('with no middleware', () => {
     it('returns the global fetch function', () => {
-      expect(createFetch()).toBe(globalFetch)
+      expect(createFetch()).toBe(global.fetch)
     })
   })
 })