+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## HEAD
 
 - Rename `onResponse` to `recv` and `enhanceFetch` to `enableRecv`. This makes the association more clear.
+- Deprecated redundant top-level `fetch` export. Use the global `fetch` function directly instead.
 
 ## [v4.2.0]
 > Sep 29, 2016